Wolontariat_back_Django/requirements.txt
Django~=5.0
django-cors-headers~=4.3.1
djangorestframework~=3.14.0
psycopg2~=2.9.9
python-dotenv~=1.0.0